Ensure that no unexpected perso is loaded
This change ensures that loading the simulator is aborted if the
indicated perso can not be loaded. This behavior is usefull to ensure
that the user does not operate on a different perso than expected. It
was default behavior up to some commit in the near past and should not
have been changed in the first place.