Privileged-only CA-Keys
Keys are assigned a privileged-only field. Usage is protocol dependent.
Implementation within AbstractCaProtocol does include privileged-only
key only within a PrivilegedTerminalInfo when SecurityInfos for
SecInfoPublicity.PRIVILEGED are requested.

Within nPA-Context these leads to inclusion of the Public key in
EF.ChipSecurity only and is thus accesible only by privileged terminals.
Other checks are thus not required as the public key is only available
to privileged terminals and without the public key any access is nearly
useless (beside some security considerations, which are not relevant in
the simulation context)