Allow access to CardObjects as long as no SecStatus available
As long as a Personalization is not initialized (e.g. the simulator
started) the SecStatus might not be propagated to the contained
CardObjects. Anyways it needs to be possible to access parts of it for
various purposes (reading for ControlInterface, creating the perso
initially). Thus all access checks are only enforced when the SecStatus
is already set.