Current file can be stored in SecStatus
In order to enable (re-)storing of session contexts it is needed that
the currently selected file is stored within the SecStatus. This commit
lays the foundation for this behavior and allows unified handling of
current file through the CurrentFileHandler.

In the following commits this will be used in the relevant protocols.