Adapt TypeIdentifier to upstream changes in CardObjectIdentifier
As a CardObjectIdentifier only needs to match on concrete objects there
is no longer a need to attach TypeIdentifiers to object to allow that
matching. Thus the field identifier can be removed form
AbstractCardObject and the implementation in TypeIdentifier can be
streamlined.

The DateTimeObjectIdentifier essentially is only a specific instance of
a TypeIdentifier and thus is replaced by the generic type.